fix: ensure authentication
Quick fix for #173. Note, now we're requesting a session twice as
ehaCouchDbAuthService's login is not working within PouchDB's context.

I expected cookie auth to work (we're not using its http interceptor to add a
bearer token as we're using PouchDB as a client); CouchDB is correctly
responding with an HttpOnly cookie, but it is not being sent in subsequent
requests.